+++ wifipumpkin3/core/packets/dnsserver.py
@@ -13,6 +13,22 @@ from dnslib.proxy import ProxyResolver
 from dnslib.server import DNSServer
 import binascii
 
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 TYPE_LOOKUP = {
     'A': (dns.A, QTYPE.A),
@@ -68,7 +84,6 @@ class Record:
 
     def __str__(self):
         return str(self.rr)
-
 
 class Resolver(ProxyResolver):
 
@@ -133,8 +148,6 @@ class Resolver(ProxyResolver):
 
         self.output.emit('no local zone found, proxying %s[%s]' % (request.q.qname, type_name))
         return super().resolve(request, handler)
-
-
 
 class LocalDNSLogger(object):
     
@@ -217,8 +230,6 @@ class LocalDNSLogger(object):
     def log_data(self, dnsobj):
         self.logger.emit("\n" + dnsobj.toZone("    ") + "\n")
 
-
-
 class DNSServerThread(QThread):
     ''' Simple DNS server UDP resolver '''
     sendRequests = pyqtSignal(object) #I'll use this object in future feature
@@ -226,7 +237,6 @@ class DNSServerThread(QThread):
         super(DNSServerThread, self).__init__(parent = None)
         self.resolver = None
         self.conf = conf
-
 
     def run(self):