+++ wifipumpkin3/core/packets/dnsspoofNF.py
@@ -5,40 +5,22 @@ logging.getLogger('scapy.runtime').setLevel(logging.ERROR)
 from scapy.all import *
 from netfilterqueue import NetfilterQueue
 
-"""
-Description:
-    This program is a module for wifi-pumpkin.py file which includes new implementation
-    for Dns spoof Attack with NetfilterQueue and iptables.
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
 
-Copyright:
-    Copyright (C) 2015-2016 Marcos Nesster P0cl4bs Team
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
 
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>
-"""
+# http://www.apache.org/licenses/LICENSE-2.0
 
-'''http://stackoverflow.com/questions/17035077/python-logging-to-multiple-log-files-from-different-classes'''
-def setup_logger(logger_name, log_file, level=logging.INFO):
-    l = logging.getLogger(logger_name)
-    formatter = logging.Formatter('%(message)s')
-    fileHandler = logging.FileHandler(log_file, mode='a')
-    fileHandler.setFormatter(formatter)
-    streamHandler = logging.StreamHandler()
-    streamHandler.setFormatter(formatter)
-
-    l.setLevel(level)
-    l.addHandler(fileHandler)
-    l.addHandler(streamHandler)
-
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 class DnsSpoofNetFilter(object):
     def __init__(self):
@@ -51,9 +33,11 @@ class DnsSpoofNetFilter(object):
         self.args = parser.parse_args()
 
     def logggingCreate(self):
-        setup_logger('dnsspoofAP', './logs/AccessPoint/dnsspoof.log')
-        self.logDNS = logging.getLogger('dnsspoofAP')
-        self.logDNS.info('Dns Spoof: running...')
+        #TODO: implement this function logger with new log implementated with json
+        pass
+        # setup_logger('dnsspoofAP', './logs/AccessPoint/dnsspoof.log')
+        # self.logDNS = logging.getLogger('dnsspoofAP')
+        # self.logDNS.info('Dns Spoof: running...')
 
     def callback(self,packet):
         payload = packet.get_payload()