+++ wifipumpkin3/core/servers/mitm/responder3.py
@@ -4,6 +4,22 @@ from wifipumpkin3.core.config.globalimport import *
 from wifipumpkin3.core.widgets.docks.dock import DockableWidget
 from wifipumpkin3.core.controls.threads import ProcessThread
 
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 class NetCredential(DockableWidget):
     id = "Responder3"
@@ -44,7 +60,7 @@ class Responder3(MitmMode):
 
     @property
     def CMD_ARRAY(self):
-        iface  = self.conf.get('accesspoint', 'interfaceAP')
+        iface  = self.conf.get('accesspoint', 'interface')
         config_responder3_path = self.conf.get('mitm_modules', 'responder3_config')
         self._cmd_array=['-I',iface , '-4', '-p', config_responder3_path]
         return self._cmd_array