+++ wifipumpkin3/core/servers/proxy/captiveflask.py
@@ -6,13 +6,29 @@ import queue
 from scapy.all import *
 import logging, os
 import wifipumpkin3.core.utility.constants as C
-from wifipumpkin3.core.common.platforms import setup_logger
 from wifipumpkin3.core.servers.proxy.proxymode import *
 from wifipumpkin3.core.utility.collection import SettingsINI
 from wifipumpkin3.core.common.uimodel import *
 from wifipumpkin3.core.widgets.docks.dock import DockableWidget
 from wifipumpkin3.plugins.captivePortal import *
 from ast import literal_eval
+
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 class TCPProxyDock(DockableWidget):
     id = "TCPProxy"
@@ -54,7 +70,6 @@ class CaptivePortal(ProxyMode):
         self.setTypePlugin(self.TypePlugin)
         self.plugins        = []
         self.search_all_ProxyPlugins()
-        #self.setRunningPort(self.conf.get('proxy_plugins', 'pumpkinproxy_config_port'))
 
     @property
     def CMD_ARRAY(self):
@@ -73,7 +88,7 @@ class CaptivePortal(ProxyMode):
         self.reactor.setObjectName(self.ID)
 
         # settings iptables for add support captive portal 
-        IFACE = self.conf.get('accesspoint', 'interfaceAP')
+        IFACE = self.conf.get('accesspoint', 'interface')
         IP_ADDRESS = self.conf.get('dhcp', 'router')
         PORT= 80
         
@@ -89,7 +104,6 @@ class CaptivePortal(ProxyMode):
         self.defaults_rules[self.ID].append('iptables -t nat -A PREROUTING -i {iface} -p tcp --dport 80 -j DNAT --to-destination {ip}:{port}'.format(iface=IFACE,ip=IP_ADDRESS, port=PORT))
         self.runDefaultRules()
 
-
     @property
     def getPlugins(self):
         commands = self.config.get_all_childname('plugins')
@@ -100,7 +114,7 @@ class CaptivePortal(ProxyMode):
 
     def LogOutput(self,data):
         headers_table, output_table = ["IP", "Login", "Password"], []
-        if self.conf.get('accesspoint', 'statusAP', format=bool):
+        if self.conf.get('accesspoint', 'status_ap', format=bool):
             self.logger.info(data)
             try:
                 data = literal_eval(data)