+++ wifipumpkin3/core/servers/proxy/pumpkin_proxy.py
@@ -6,11 +6,27 @@ import queue
 from scapy.all import *
 import logging, os
 import wifipumpkin3.core.utility.constants as C
-from wifipumpkin3.core.common.platforms import setup_logger
 from wifipumpkin3.core.servers.proxy.proxymode import *
 from wifipumpkin3.core.utility.collection import SettingsINI
 from wifipumpkin3.core.common.uimodel import *
 from wifipumpkin3.core.widgets.docks.dock import DockableWidget
+
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 class TCPProxyDock(DockableWidget):
     id = "TCPProxy"
@@ -74,7 +90,7 @@ class PumpKinProxy(ProxyMode):
         return list_commands
 
     def LogOutput(self,data):
-        if self.conf.get('accesspoint', 'statusAP', format=bool):
+        if self.conf.get('accesspoint', 'status_ap', format=bool):
             self.logger.info(data)
 
     def parser_set_pumpkinproxy(self, status, plugin_name):