+++ wifipumpkin3/core/widgets/default/session_config.py
@@ -2,6 +2,23 @@ from wifipumpkin3.core.config.globalimport import *
 from wifipumpkin3.core.common.uimodel import *
 import weakref
 
+# This file is part of the wifipumpkin3 Open Source Project.
+# wifipumpkin3 is licensed under the Apache 2.0.
+
+# Copyright 2020 P0cL4bs Team - Marcos Bomfim (mh4x0f)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+# http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 class SessionConfig(TabsWidget):
     ConfigRoot="Settings"
     Name = "Settings"
@@ -20,10 +37,7 @@ class SessionConfig(TabsWidget):
         self.settingsItem = {}
 
         for mod in settingsItem:
-            #print(mod.title)
             self.settingsItem[mod.title]=mod
-            #self.mainlayout.addWidget(mod)
-            #Hack to add all the modules into class
             setattr(self.__class__,mod.ID,mod)
 
     @property